Use new AI for hologrems
