+++ Assets/Unity/Animations/Training Game/Hologrem/Distortion.anim
@@ -35,7 +35,7 @@ AnimationClip:
       value: {fileID: -2541546423034188577, guid: 8a82eb088810d459580da49adf029e8c,
         type: 3}
     attribute: m_Sprite
-    path: 
+    path: Sprite
     classID: 212
     script: {fileID: 0}
   m_SampleRate: 60
