--- Assets/src/MoveTowardsPlayerAI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ae823e05fdfae40fa8573f065f82a88f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
