--- Assets/src/PassiveSounds.cs
@@ -1,50 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class PassiveSounds : MonoBehaviour {
-  public GameObject EnemyGameObject;
-  // public SoundController SoundPlayer;
-  public List<AudioClip> Sounds;
-  public float MinInterval, MaxInterval;
-
-  private IntervalTimer Timer;
-
-  void Start() {
-    Timer = new IntervalTimer() {
-      TimeClass = "PlayingTime",
-    };
-
-    UpdatePlayInterval();
-  }
-
-  void Update() {
-    if (!StateManager.Playing)
-      return;
-
-    Timer.IfElapsed(PlaySound);
-  }
-
-  void PlaySound() {
-    if (EnemyUtils.IsDead(EnemyGameObject))
-      return;
-
-    // Because C# doesn't appear to have any built-in 
-    // random sampling methods. 
-    int i = UnityEngine.Random.Range(0, Sounds.Count);
-    var sound = Sounds[i];
-    Debug.LogError("PassiveSounds is using deprecated SoundController");
-    // SoundPlayer.Play(sound);
-
-    UpdatePlayInterval();
-  }
-
-  void UpdatePlayInterval() {
-    Timer.Interval = MinInterval
-      + (MaxInterval - MinInterval)
-      * (UnityEngine.Random.value);
-
-    Timer.Reset();
-  }
-}
