--- Assets/src/PassiveSounds.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 76d7bdd39ad3d4b5a893da647ece1e29
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
